requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
click==8.1.7
decorator==5.1.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
flake8==7.0.0
ipython==8.24.0
isort==5.13.2
jedi==0.19.1
Markdown==3.6
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
